+++ formspree/__init__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
-import app
-forms_app = app.app
+
+from app import create_app
+forms_app = create_app()
 
 if __name__ == '__main__':
     forms_app.run(host='0.0.0.0')
