+++ formspree/static/scss/ionicons.min.scss
